+++ packages/amplify-category-api/CHANGELOG.md
@@ -3,6 +3,14 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+## [3.1.3](https://github.com/aws-amplify/amplify-category-api/compare/@aws-amplify/amplify-category-api@3.1.2...@aws-amplify/amplify-category-api@3.1.3) (2022-07-01)
+
+**Note:** Version bump only for package @aws-amplify/amplify-category-api
+
+
+
+
+
 ## [3.1.2](https://github.com/aws-amplify/amplify-category-api/compare/@aws-amplify/amplify-category-api@3.1.1...@aws-amplify/amplify-category-api@3.1.2) (2022-06-23)
 
 **Note:** Version bump only for package @aws-amplify/amplify-category-api