+++ packages/amplify-graphql-maps-to-transformer/CHANGELOG.md
@@ -3,6 +3,14 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+## [1.1.21](https://github.com/aws-amplify/amplify-category-api/compare/@aws-amplify/graphql-maps-to-transformer@1.1.20...@aws-amplify/graphql-maps-to-transformer@1.1.21) (2022-07-01)
+
+**Note:** Version bump only for package @aws-amplify/graphql-maps-to-transformer
+
+
+
+
+
 ## [1.1.20](https://github.com/aws-amplify/amplify-category-api/compare/@aws-amplify/graphql-maps-to-transformer@1.1.19...@aws-amplify/graphql-maps-to-transformer@1.1.20) (2022-06-23)
 
 **Note:** Version bump only for package @aws-amplify/graphql-maps-to-transformer