+++ packages/graphql-relational-schema-transformer/CHANGELOG.md
@@ -3,6 +3,14 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+## [2.21.10](https://github.com/aws-amplify/amplify-category-api/compare/graphql-relational-schema-transformer@2.21.9...graphql-relational-schema-transformer@2.21.10) (2022-07-01)
+
+**Note:** Version bump only for package graphql-relational-schema-transformer
+
+
+
+
+
 ## [2.21.9](https://github.com/aws-amplify/amplify-category-api/compare/graphql-relational-schema-transformer@2.21.8...graphql-relational-schema-transformer@2.21.9) (2022-06-23)
 
 **Note:** Version bump only for package graphql-relational-schema-transformer